Rename test_groups to test_group
